Bugfix
Found error with using ';' as the end index of JSON object when a string
in the object contained ';'. Prior, an arbitrary end index for the JSON
string was used due to a perceived efficiency error with larger objects;
however, after realizing the source of the error, the entire object is
now used in the string.